# frontend/teampilot/openapi.yaml
openapi: 3.0.1
info:
  title: TeamPilot.Api
  version: '1.0'
paths:
...
  /users/{id}:
    get:
      tags:
        - User
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/UserWithNullablePlayerPropsDTO'
            application/json:
              schema:
                $ref: '#/components/schemas/UserWithNullablePlayerPropsDTO'
            text/json:
              schema:
                $ref: '#/components/schemas/UserWithNullablePlayerPropsDTO'
    put:
      tags:
        - User
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserWithNullablePlayerPropsDTO'
          text/json:
            schema:
              $ref: '#/components/schemas/UserWithNullablePlayerPropsDTO'
          application/*+json:
            schema:
              $ref: '#/components/schemas/UserWithNullablePlayerPropsDTO'
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/UserWithNullablePlayerPropsDTO'
            application/json:
              schema:
                $ref: '#/components/schemas/UserWithNullablePlayerPropsDTO'
            text/json:
              schema:
                $ref: '#/components/schemas/UserWithNullablePlayerPropsDTO'
    delete:
      tags:
        - User
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
  /users/player/tournaments-and-matches/{playerId}:
    get:
      tags:
        - User
      parameters:
        - name: playerId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TournamentMatchDTO'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TournamentMatchDTO'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TournamentMatchDTO'
  /users/tournament-manager/tournaments-and-matches/{tournamentManagerId}:
    get:
      tags:
        - User
      parameters:
        - name: tournamentManagerId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TournamentMatchDTO'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TournamentMatchDTO'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TournamentMatchDTO'
  /users/team-manager/tournaments-and-matches/{teamManagerId}:
    get:
      tags:
        - User
      parameters:
        - name: teamManagerId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TournamentMatchDTO'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TournamentMatchDTO'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TournamentMatchDTO'
  /users/{userId}/follow/{itemType}/{itemId}:
    post:
      tags:
        - User
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: itemId
          in: path
          required: true
          schema:
            type: string
        - name: itemType
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
  /users/{userId}/unfollow/{itemType}/{itemId}:
    delete:
      tags:
        - User
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: itemId
          in: path
          required: true
          schema:
            type: string
        - name: itemType
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
  /users/{userId}/unfollowAll:
    delete:
      tags:
        - User
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
components:
  schemas:
...
    MatchDTO:
      type: object
      properties:
        matchId:
          type: string
          nullable: true
        matchLengthInMinutes:
          type: string
          nullable: true
        matchDate:
          type: string
          nullable: true
        participatingTeams:
          type: array
          items:
            $ref: '#/components/schemas/TeamForTournamentDTO'
          nullable: true
      additionalProperties: false
...
    PastMatchDTO:
      type: object
      properties:
        matchId:
          type: string
          nullable: true
        matchLengthInMinutes:
          type: string
          nullable: true
        matchDate:
          type: string
          nullable: true
        participatingTeams:
          type: array
          items:
            $ref: '#/components/schemas/TeamForTournamentDTO'
          nullable: true
        winningTeam:
          $ref: '#/components/schemas/TeamForTournamentDTO'
      additionalProperties: false
...
    TeamForTournamentDTO:
      type: object
      properties:
        teamId:
          type: string
          nullable: true
        teamName:
          type: string
          nullable: true
        avatarURL:
          type: string
          nullable: true
      additionalProperties: false
    TournamentDTO:
      type: object
      properties:
        tournamentId:
          type: string
          nullable: true
        tournamentName:
          type: string
          nullable: true
        tournamentFormat:
          type: string
          nullable: true
        tournamentPrizePool:
          type: number
          format: double
        tournamentLocation:
          type: string
          nullable: true
        tournamentSponsor:
          type: string
          nullable: true
        tournamentStartDate:
          type: string
          nullable: true
        tournamentEndDate:
          type: string
          nullable: true
        tournamentMatches:
          type: array
          items:
            $ref: '#/components/schemas/MatchDTO'
          nullable: true
        futureTournamentMatches:
          type: array
          items:
            $ref: '#/components/schemas/MatchDTO'
          nullable: true
        pastTournamentMatches:
          type: array
          items:
            $ref: '#/components/schemas/PastMatchDTO'
          nullable: true
        tournamentParticipatingTeams:
          type: array
          items:
            $ref: '#/components/schemas/TeamForTournamentDTO'
          nullable: true
        winningTeam:
          $ref: '#/components/schemas/TeamForTournamentDTO'
        upcomingMatch:
          $ref: '#/components/schemas/MatchDTO'
      additionalProperties: false
...
    TournamentMatchDTO:
      type: object
      properties:
        tournament:
          $ref: '#/components/schemas/TournamentDTO'
        match:
          $ref: '#/components/schemas/MatchDTO'
      additionalProperties: false
...
    UserWithNullablePlayerPropsDTO:
      type: object
      properties:
        userId:
          type: string
          nullable: true
        userType:
          type: string
          nullable: true
        nickname:
          type: string
          nullable: true
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        monthlySalary:
          type: number
          format: double
          nullable: true
        teamId:
          type: string
          nullable: true
      additionalProperties: false